Commit : - Ajout de logos dans la barre d'outils ainsi que tous les boutons rattachés

+++ microsoftears/src/main/java/ca/ulaval/glo2004/utilitaires/Pouce.java
@@ -221,9 +221,10 @@ public class Pouce
         return new Pouce(pouces, numerateur, denominateur);
     }
 
-    // TODO : à vérifier car int -> division float
+    // TODO : à vérifier
     public double toDouble(){
-        return this.getPouces()+(this.getNumerateur()/this.getDenominateur());
+
+        return (double) this.getPouces() + this.getNumerateur() / this.getDenominateur();
     }
 
     /** Calcul du plus grand dénominateur commun (pgdc)
